+++ blockchain/hot/pool_test.go
@@ -1,6 +1,7 @@
 package hot
 
 import (
+	"github.com/tendermint/tendermint/store"
 	"os"
 	"sort"
 	"testing"
@@ -10,7 +11,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	abci "github.com/tendermint/tendermint/abci/types"
-	"github.com/tendermint/tendermint/blockchain"
 	cfg "github.com/tendermint/tendermint/config"
 	cmn "github.com/tendermint/tendermint/libs/common"
 	dbm "github.com/tendermint/tendermint/libs/db"
@@ -504,7 +504,7 @@ func newBlockchainPoolPair(logger log.Logger, genDoc *types.GenesisDoc, privVals
 	}
 	blockDB := dbm.NewMemDB()
 	stateDB := dbm.NewMemDB()
-	blockStore := blockchain.NewBlockStore(blockDB)
+	blockStore := store.NewBlockStore(blockDB)
 	state, err := sm.LoadStateFromDBOrGenesisDoc(stateDB, genDoc)
 	if err != nil {
 		panic(cmn.ErrorWrap(err, "error constructing state from genesis file"))