+++ node/node.go
@@ -253,13 +253,16 @@ func createAndStartIndexerService(config *cfg.Config, dbProvider DBProvider,
 		if err != nil {
 			return
 		}
+		indexOptions := make([]func(index *kv.TxIndex), 0)
 		if config.TxIndex.IndexTags != "" {
-			txIndexer = kv.NewTxIndex(store, kv.IndexTags(splitAndTrimEmpty(config.TxIndex.IndexTags, ",", " ")))
+			indexOptions = append(indexOptions, kv.IndexTags(splitAndTrimEmpty(config.TxIndex.IndexTags, ",", " ")))
 		} else if config.TxIndex.IndexAllTags {
-			txIndexer = kv.NewTxIndex(store, kv.IndexAllTags())
-		} else {
-			txIndexer = kv.NewTxIndex(store)
+			indexOptions = append(indexOptions, kv.IndexAllTags())
 		}
+		if config.TxIndex.EnableRangeQuery {
+			indexOptions = append(indexOptions, kv.EnableRangeQuery())
+		}
+		txIndexer = kv.NewTxIndex(store, indexOptions...)
 	default:
 		txIndexer = &null.TxIndex{}
 	}
@@ -963,7 +966,11 @@ func (n *Node) startRPC() ([]net.Listener, error) {
 				}
 			}), rpcserver.SetWorkerPool(wsWorkerPool), rpcserver.ReadLimit(config.MaxBodyBytes))
 		wm.SetLogger(wmLogger)
-		mux.HandleFunc("/websocket", wm.WebsocketHandler)
+		if n.config.RPC.DisableWebsocket {
+			mux.HandleFunc("/websocket", wm.WebsocketDisabledHandler)
+		} else {
+			mux.HandleFunc("/websocket", wm.WebsocketHandler)
+		}
 		rpcserver.RegisterRPCFuncs(mux, rpccore.Routes, coreCodec, rpcLogger)
 		listener, err := rpcserver.Listen(
 			listenAddr,