+++ state/txindex/kv/kv_test.go
@@ -10,11 +10,7 @@ import (
 	"github.com/stretchr/testify/require"
 	abci "github.com/tendermint/tendermint/abci/types"
 	cmn "github.com/tendermint/tendermint/libs/common"
-	db "github.com/tendermint/tendermint/libs/db"
-
-	abci "github.com/tendermint/tendermint/abci/types"
-	cmn "github.com/tendermint/tendermint/libs/common"
-	"github.com/tendermint/tendermint/libs/pubsub/query"
+	"github.com/tendermint/tendermint/libs/db"
 	"github.com/tendermint/tendermint/state/txindex"
 	"github.com/tendermint/tendermint/types"
 )
@@ -67,7 +63,7 @@ func TestTxIndex(t *testing.T) {
 
 func TestTxSearch(t *testing.T) {
 	allowedTags := []string{"account.number", "account.owner", "account.date"}
-	indexer := NewTxIndex(db.NewMemDB(), IndexTags(allowedTags))
+	indexer := NewTxIndex(db.NewMemDB(), IndexTags(allowedTags), EnableRangeQuery())
 
 	txResult := txResultWithEvents([]abci.Event{
 		{Type: "account", Attributes: []cmn.KVPair{{Key: []byte("number"), Value: []byte("1")}}},
@@ -120,7 +116,7 @@ func TestTxSearch(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.q, func(t *testing.T) {
-			results, err := indexer.Search(query.MustParse(tc.q))
+			results, err := indexer.Search(tc.q)
 			assert.NoError(t, err)
 
 			assert.Len(t, results, tc.resultsLength)
@@ -193,7 +189,7 @@ func TestTxSearchDeprecatedIndexing(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.q, func(t *testing.T) {
-			results, err := indexer.Search(query.MustParse(tc.q))
+			results, err := indexer.Search(tc.q)
 			require.NoError(t, err)
 			require.Equal(t, results, tc.results)
 		})
@@ -202,7 +198,7 @@ func TestTxSearchDeprecatedIndexing(t *testing.T) {
 
 func TestTxSearchOneTxWithMultipleSameTagsButDifferentValues(t *testing.T) {
 	allowedTags := []string{"account.number"}
-	indexer := NewTxIndex(db.NewMemDB(), IndexTags(allowedTags))
+	indexer := NewTxIndex(db.NewMemDB(), IndexTags(allowedTags), EnableRangeQuery())
 
 	txResult := txResultWithEvents([]abci.Event{
 		{Type: "account", Attributes: []cmn.KVPair{{Key: []byte("number"), Value: []byte("1")}}},
@@ -212,7 +208,7 @@ func TestTxSearchOneTxWithMultipleSameTagsButDifferentValues(t *testing.T) {
 	err := indexer.Index(txResult)
 	require.NoError(t, err)
 
-	results, err := indexer.Search(query.MustParse("account.number >= 1"))
+	results, err := indexer.Search("account.number >= 1")
 	assert.NoError(t, err)
 
 	assert.Len(t, results, 1)
@@ -221,7 +217,7 @@ func TestTxSearchOneTxWithMultipleSameTagsButDifferentValues(t *testing.T) {
 
 func TestTxSearchMultipleTxs(t *testing.T) {
 	allowedTags := []string{"account.number", "account.number.id"}
-	indexer := NewTxIndex(db.NewMemDB(), IndexTags(allowedTags))
+	indexer := NewTxIndex(db.NewMemDB(), IndexTags(allowedTags), EnableRangeQuery())
 
 	// indexed first, but bigger height (to test the order of transactions)
 	txResult := txResultWithEvents([]abci.Event{
@@ -266,7 +262,7 @@ func TestTxSearchMultipleTxs(t *testing.T) {
 	err = indexer.Index(txResult4)
 	require.NoError(t, err)
 
-	results, err := indexer.Search(query.MustParse("account.number >= 1"))
+	results, err := indexer.Search("account.number >= 1")
 	assert.NoError(t, err)
 
 	require.Len(t, results, 3)
@@ -274,7 +270,7 @@ func TestTxSearchMultipleTxs(t *testing.T) {
 }
 
 func TestIndexAllTags(t *testing.T) {
-	indexer := NewTxIndex(db.NewMemDB(), IndexAllTags())
+	indexer := NewTxIndex(db.NewMemDB(), IndexAllTags(), EnableRangeQuery())
 
 	txResult := txResultWithEvents([]abci.Event{
 		{Type: "account", Attributes: []cmn.KVPair{{Key: []byte("owner"), Value: []byte("Ivan")}}},
@@ -284,15 +280,25 @@ func TestIndexAllTags(t *testing.T) {
 	err := indexer.Index(txResult)
 	require.NoError(t, err)
 
-	results, err := indexer.Search(query.MustParse("account.number >= 1"))
+	results, err := indexer.Search("account.number >= 1")
 	assert.NoError(t, err)
 	assert.Len(t, results, 1)
 	assert.Equal(t, []*types.TxResult{txResult}, results)
 
-	results, err = indexer.Search(query.MustParse("account.owner = 'Ivan'"))
+	results, err = indexer.Search("account.owner = 'Ivan'")
 	assert.NoError(t, err)
 	assert.Len(t, results, 1)
 	assert.Equal(t, []*types.TxResult{txResult}, results)
+}
+
+
+func TestDisableRangeQuery(t *testing.T) {
+	indexer := NewTxIndex(db.NewMemDB(), IndexAllTags())
+
+	_, err := indexer.Search("account.number >= 1")
+	assert.Error(t, err)
+	_, err = indexer.Search("account.number >= 1 AND account.sequence < 100 AND tx.height > 200 AND tx.height <= 300")
+	assert.Error(t, err)
 }
 
 func txResultWithEvents(events []abci.Event) *types.TxResult {